+++ packages/panels/src/almasix/orbit/panels/__init__.py
@@ -1,5 +1,5 @@
 from almasix.orbit.panels.auth import Login, Register
-from almasix.orbit.panels.navigation import NavigationGroup, NavigationItem
+from almasix.orbit.panels.navigation import NavigationGroup, NavigationItem, NavigationSubgroup
 from almasix.orbit.panels.page import Page
 from almasix.orbit.panels.pages.dashboard import Dashboard
 from almasix.orbit.panels.panel import Panel, PanelRegistry
@@ -18,6 +18,7 @@ __all__ = [
     "RelationManager",
     "NavigationItem",
     "NavigationGroup",
+    "NavigationSubgroup",
     "OrbitUser",
     "UserMenuItem",
     "PanelNotification",